feat(MovieRepositoryImpl): 데이터 계층의 레포지토리 구현체를 임시로 프레젠테이션 계층에서 구현
